refactor(rows): split repo.rs + rest.rs into focused modules
repo.rs (1620 lines) → repo/ module:
  - users.rs      — login, register, sessions, admin
  - characters.rs — character CRUD, stats, positions, custom data
  - instances.rs  — zone instances, world servers, allocation
  - global_data.rs — key-value store
  - abilities.rs  — ability CRUD, bars
  - zones.rs      — zone management

rest.rs (1426 lines) → rest/ module:
  - auth.rs       — public API (login, register, character ops)
  - instances.rs  — instance management routes
  - characters.rs — character persistence routes
  - abilities.rs  — ability routes
  - zones.rs      — zone management routes
  - global_data.rs — global data routes
  - management.rs — admin user CRUD

No logic changes — mechanical split only. Each file < 550 lines.
Enables targeted optimization with &str refs per module.
